# outputs/english-platform-v1-1/openapi-v1.yaml
openapi: 3.1.0
...
info:
  title: 个性化英语学习平台 API
  version: 1.1.0
  summary: General 与 TOEFL 核心学习闭环的官方 Web API 契约
  description: |
    本契约是 V1.1 服务端与官方 Web 端之间的唯一接口真相源。

    所有租户资源均位于 `/api/v1/tenants/{tenantId}/...`。服务端必须先验证当前
    用户的有效租户成员关系，再在数据库事务内设置租户、用户和成员上下文，由
    PostgreSQL RLS 执行第二道隔离。调用者无权知道目标资源是否存在时统一返回
    `404`；资源本身对调用者可见、但当前操作不被允许时才返回 `403`。

    日期时间字段必须是 UTC RFC 3339（以 `Z` 结尾），标识符均为 UUID。列表使用
    不透明游标分页。错误响应使用 RFC 7807 `application/problem+json`。

    官方 Web 端使用 HttpOnly Access Cookie。所有有状态写请求还必须携带
    `X-CSRF-Token`，其值与可读 CSRF Cookie 组成双提交校验。标记了
    `Idempotency-Key` 的命令在相同主体、租户、路由和请求体下重试时必须返回首次
    结果；同键不同请求体返回 `409`。自动保存使用 `If-Match` 与响应 `ETag`，并以
    `revision` 提供可观测的乐观并发版本。
  contact:
    name: English Platform Engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Same-origin official Web application
...
security:
  - accessCookie: []
paths:
...
  /api/v1/tenants/{tenantId}/admin/memberships:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    get:
      tags: [Admin Memberships]
      summary: List tenant memberships
      operationId: listMemberships
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/TenantRole'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/MembershipStatus'
      responses:
        '200':
          description: Tenant memberships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Admin Memberships]
      summary: Create a tenant membership and its role rows
      operationId: createMembership
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembershipRequest'
      responses:
        '201':
          description: Membership created
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/memberships/{membershipId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/MembershipId'
    patch:
      tags: [Admin Memberships]
      summary: Replace roles or change membership lifecycle status
      operationId: updateMembership
      security:
        - accessCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMembershipRequest'
      responses:
        '200':
          description: Updated membership
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/memberships/{membershipId}/reset-password:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/MembershipId'
    post:
      tags: [Admin Memberships]
      summary: Reset a student's password and revoke active sessions
      operationId: resetStudentPassword
      security:
        - accessCookie: []
          csrfHeader: []
      responses:
        '200':
          description: One-time temporary credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentCredential'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/tenants/{tenantId}/admin/student-accounts:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    post:
      tags: [Admin Memberships]
      summary: Provision one active student account
      operationId: createStudentAccount
      security:
        - accessCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentAccountInput'
      responses:
        '200':
          description: One-time temporary credential
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/StudentCredential'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/student-accounts/bulk:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    post:
      tags: [Admin Memberships]
      summary: Provision up to 200 active student accounts atomically
      operationId: createStudentAccounts
      security:
        - accessCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [students]
              properties:
                students:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items:
                    $ref: '#/components/schemas/StudentAccountInput'
      responses:
        '200':
          description: One-time temporary credentials
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentCredential'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/contents:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    get:
      tags: [Admin Content]
      summary: List tenant-owned content and official content from the separate platform schema
      operationId: listContents
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: ownership
          in: query
          schema:
            $ref: '#/components/schemas/CatalogOwnership'
        - name: kind
          in: query
          schema:
            $ref: '#/components/schemas/ContentKind'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/EntityStatus'
      responses:
        '200':
          description: Official objects from the separate platform schema are exposed as read-only projections.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Admin Content]
      summary: Create tenant content or clone an official version from the separate platform schema
      operationId: createContent
      description: Supplying cloneFromPlatformVersionId copies the dependency graph into a new tenant-owned entity and version; the separate platform schema is never modified.
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentRequest'
      responses:
        '201':
          description: Tenant-owned content created
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/contents/{contentId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/ContentId'
    get:
      tags: [Admin Content]
      summary: Get a content entity and version metadata
      operationId: getContent
      responses:
        '200':
          description: Content entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/tenants/{tenantId}/admin/contents/{contentId}/versions:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/ContentId'
    post:
      tags: [Admin Content]
      summary: Create an immutable content draft version
      operationId: createContentVersion
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentVersionRequest'
      responses:
        '201':
          description: Draft version created
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentVersion'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/content-versions/{versionId}/publish:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/VersionId'
    post:
      tags: [Admin Content]
      summary: Publish a content version immutably
      operationId: publishContentVersion
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Published version
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/questions:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    get:
      tags: [Admin Questions]
      summary: List tenant-owned questions and official questions from the separate platform schema
      operationId: listQuestions
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: ownership
          in: query
          schema:
            $ref: '#/components/schemas/CatalogOwnership'
        - name: kind
          in: query
          schema:
            $ref: '#/components/schemas/QuestionKind'
      responses:
        '200':
          description: Official question projections from the separate platform schema are read-only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Admin Questions]
      summary: Create a tenant question or clone an official version from the separate platform schema
      operationId: createQuestion
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionRequest'
      responses:
        '201':
          description: Tenant-owned question created
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/questions/{questionId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/QuestionId'
    get:
      tags: [Admin Questions]
      summary: Get a question entity and version metadata
      operationId: getQuestion
      responses:
        '200':
          description: Question entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/tenants/{tenantId}/admin/questions/{questionId}/versions:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/QuestionId'
    post:
      tags: [Admin Questions]
      summary: Create an immutable question draft version
      operationId: createQuestionVersion
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionVersionRequest'
      responses:
        '201':
          description: Question draft version created
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionVersionAdmin'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/question-versions/{versionId}/publish:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/VersionId'
    post:
      tags: [Admin Questions]
      summary: Publish a question version immutably
      operationId: publishQuestionVersion
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Published question version
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionVersionAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tenants/{tenantId}/admin/tasks:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    get:
      tags: [Admin Tasks]
      summary: List tenant-owned tasks and official tasks from the separate platform schema
      operationId: listTasks
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: ownership
          in: query
          schema:
            $ref: '#/components/schemas/CatalogOwnership'
      responses:
        '200':
          description: Task catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDefinitionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Admin Tasks]
      summary: Create a tenant task or clone an official version from the separate platform schema
      operationId: createTaskDefinition
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskDefinitionRequest'
      responses:
        '201':
          description: Tenant-owned task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDefinition'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/tenants/{tenantId}/admin/task-versions/{versionId}/publish:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/VersionId'
    post:
      tags: [Admin Tasks]
      summary: Publish a task version immutably
      operationId: publishTaskVersion
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Published task version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskVersionAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: Short-lived, HttpOnly, Secure, SameSite=Lax signed access cookie.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotating, HttpOnly, Secure, SameSite=Strict opaque refresh cookie; only its hash is stored.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit token that must match the readable csrf_token cookie on every state-changing request.
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      description: Tenant UUID from the current user's active membership list.
      schema:
        $ref: '#/components/schemas/Uuid'
...
  schemas:
...
    TenantRole:
      type: string
      enum: [owner, admin, teacher, student, content_editor, analyst]
...